Extract JwkUtil
